Move antehandler to its own folder
